requirements-titok.txt
torch>=2.0.0
torchvision
omegaconf
transformers
timm
open_clip_torch
einops
scipy
pillow
accelerate
gdown
huggingface-hub
einops
torch-fidelity
torchinfo
webdataset